"SpatialStratifiedSample"-class {soilcarbon}	R Documentation

A class for stratified point samples

Description

A class containing input and output maps generated by running the stratification::strata.LH method on a raster map.

Slots

variable:

object of class "character"; variable name

locations:

object of class "SpatialPoints" optimized locations

strata:

object of class "SpatialPixelsDataFrame"; contains the output stratification and probabilities

LH:

object of class "data.frame"; contains the outputs of analysis produced by the stratification package

eval:

object of class "data.frame"; contains a list of total variances

Methods

plot

signature(x = "SpatialStratifiedSample"): plots map showing the samples and the optimisation curve

Author(s)

Ichsani Wheeler

See Also

eval.LH, stratification::strata.LH

[Package soilcarbon version 0.1-0 Index]