eval.LH {soilcarbon}	R Documentation

Evaluates multiple possible stratified designs

Description

Tests multiple possible stratified designs and selects the one with the lowest total variance. Looks at different combinations of strata number, breaks and sample allocations.

Usage

eval.LH(obj, tvar = names(obj)[1], n, Ls = floor(n/2), pprob = 1)

Arguments

obj
object of type "SpatialGridDataFrame"

tvar
character; target variable name

n
integer; number of samples

Ls
integer; maximum number of strata

pprob
numeric in the range 0-1; prior probability can be set also manually

Value

Object of type "SpatialStratifiedSample".

Author(s)

Ichsani Wheeler

See Also

stratification::strata.LH

Examples


# Load the data:
data(springf)
gridded(springf) <- ~x+y
proj4string(springf) <- CRS("+init=epsg:28355")
smp <- eval.LH(springf, n=48, det.lim=.3, Ls=4)
plot(smp)
library(plotKML)
# create a new object for plotting in Google Earth:
sp.domain <- as.SpatialPolygons.SpatialPixels(smp@strata["strata"])
sp.domain <- SpatialPolygonsDataFrame(sp.domain, data.frame(ID=smp@strata$strata), match.ID = FALSE)
smp.ssp <- new("SpatialSamplingPattern", method = "Generalized Lavallee-Hidiroglou Method", 
    pattern = smp@locations, sp.domain = sp.domain)
# the same plot now in Google Earth:
shape = "http://maps.google.com/mapfiles/kml/pal2/icon18.png"
plotKML(smp.ssp, shape = shape)

[Package soilcarbon version 0.1-0 Index]