springf {soilcarbon}	R Documentation

Springfield

Description

A gridded map showing predicted soil organic carbon stock for a study area in NSW, Australia.

Usage

data(springf)

Format

Data frame; contains the following columns:

SOC

numeric; soil organic carbon in kilograms per cubic meter

x

numeric; x-coordinate in GDA94 / MGA zone 55

y

numeric; y-coordinate in GDA94 / MGA zone 55

Note

Predictions of organic carbon were produced using a regression modeling only.

Author(s)

Ichsani Wheeler

Examples


# Load the data:
data(springf)
gridded(springf) <- ~x+y
proj4string(springf) <- CRS("+init=epsg:28355")
library(plotKML)
data(R_pal)
spplot(springf, col.regions=R_pal[["soc_pal"]])

[Package soilcarbon version 0.1-0 Index]